refactor: move storage module into logstore
