+++ Cargo.toml
@@ -29,7 +29,6 @@ debug = "line-tables-only"
 delta_kernel = { version = "0.9.0", features = [
     "arrow_54",
     "developer-visibility",
-    "default-engine-rustls",
 ] }
 
 # arrow
@@ -63,7 +62,6 @@ datafusion-sql = "46"
 serde = { version = "1.0.194", features = ["derive"] }
 serde_json = "1"
 strum = { version = "0.26" }
-
 
 # "stdlib"
 bytes = { version = "1" }
