+++ crates/core/src/logstore/storage/retry_ext.rs
@@ -1,7 +1,12 @@
 //! Retry extension for [`ObjectStore`]
 
-use object_store::{path::Path, Error, ObjectStore, PutPayload, PutResult, Result};
+use ::object_store::path::Path;
+use ::object_store::{Error, ObjectStore, PutPayload, PutResult, Result};
+use humantime::parse_duration;
 use tracing::log::*;
+
+use super::StorageOptions;
+use crate::{DeltaResult, DeltaTableError};
 
 /// Retry extension for [`ObjectStore`]
 ///
@@ -73,3 +78,77 @@ pub trait ObjectStoreRetryExt: ObjectStore {
 }
 
 impl<T: ObjectStore + ?Sized> ObjectStoreRetryExt for T {}
+
+#[cfg(feature = "cloud")]
+pub trait RetryConfigParse {
+    fn parse_retry_config(
+        &self,
+        options: &StorageOptions,
+    ) -> DeltaResult<::object_store::RetryConfig> {
+        let mut retry_config = ::object_store::RetryConfig::default();
+        if let Some(max_retries) = options.0.get("max_retries") {
+            retry_config.max_retries = max_retries
+                .parse::<usize>()
+                .map_err(|e| DeltaTableError::generic(e.to_string()))?;
+        }
+
+        if let Some(retry_timeout) = options.0.get("retry_timeout") {
+            retry_config.retry_timeout = parse_duration(retry_timeout).map_err(|_| {
+                DeltaTableError::generic(format!("failed to parse \"{retry_timeout}\" as Duration"))
+            })?;
+        }
+
+        if let Some(bc_init_backoff) = options.0.get("backoff_config.init_backoff") {
+            retry_config.backoff.init_backoff = parse_duration(bc_init_backoff).map_err(|_| {
+                DeltaTableError::generic(format!(
+                    "failed to parse \"{bc_init_backoff}\" as Duration"
+                ))
+            })?;
+        }
+
+        if let Some(bc_max_backoff) = options.0.get("backoff_config.max_backoff") {
+            retry_config.backoff.max_backoff = parse_duration(bc_max_backoff).map_err(|_| {
+                DeltaTableError::generic(format!(
+                    "failed to parse \"{bc_max_backoff}\" as Duration"
+                ))
+            })?;
+        }
+
+        if let Some(bc_base) = options.0.get("backoff_config.base") {
+            retry_config.backoff.base = bc_base
+                .parse::<f64>()
+                .map_err(|e| DeltaTableError::generic(e.to_string()))?;
+        }
+
+        Ok(retry_config)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use std::time::Duration;
+
+    #[test]
+    fn test_retry_config_from_options() {
+        struct TestFactory {}
+        impl RetryConfigParse for TestFactory {}
+
+        let options = maplit::hashmap! {
+            "max_retries".to_string() => "100".to_string() ,
+            "retry_timeout".to_string()  => "300s".to_string() ,
+            "backoff_config.init_backoff".to_string()  => "20s".to_string() ,
+            "backoff_config.max_backoff".to_string()  => "1h".to_string() ,
+            "backoff_config.base".to_string()  =>  "50.0".to_string() ,
+        };
+        let retry_config = TestFactory {}
+            .parse_retry_config(&StorageOptions(options))
+            .unwrap();
+
+        assert_eq!(retry_config.max_retries, 100);
+        assert_eq!(retry_config.retry_timeout, Duration::from_secs(300));
+        assert_eq!(retry_config.backoff.init_backoff, Duration::from_secs(20));
+        assert_eq!(retry_config.backoff.max_backoff, Duration::from_secs(3600));
+        assert_eq!(retry_config.backoff.base, 50_f64);
+    }
+}
