+++ crates/core/src/operations/update.rs
@@ -1203,7 +1203,7 @@ mod tests {
 
         // Too close for missiles, switching to guns. Just checking that the data wasn't actually
         // written instead!
-        if let Ok(files) = crate::storage::utils::flatten_list_stream(
+        if let Ok(files) = crate::logstore::tests::flatten_list_stream(
             &table.object_store(),
             Some(&object_store::path::Path::from("_change_data")),
         )
