+++ crates/core/tests/command_restore.rs
@@ -3,8 +3,8 @@ use arrow_array::{Int32Array, RecordBatch};
 use arrow_schema::{DataType as ArrowDataType, Field};
 use chrono::DateTime;
 use deltalake_core::kernel::{DataType, PrimitiveType, StructField};
+use deltalake_core::logstore::commit_uri_from_version;
 use deltalake_core::protocol::SaveMode;
-use deltalake_core::storage::commit_uri_from_version;
 use deltalake_core::{DeltaOps, DeltaTable};
 use itertools::Itertools;
 use rand::Rng;
